사용하지 않는 코드들을 삭제하라
기존에 주석처리하거나 사용하지않는 코드들을 지금 사용하지않아 혼란을 줄 수 있어 삭제합니다.
