+++ app-web/src/App.tsx
@@ -9,7 +9,7 @@ import { useAppDispatch } from './hooks';
 import { setAccessToken } from './authSlice';
 
 import HeaderContainer from './HeaderContainer';
-import Seats from './seats';
+import SeatsContainer from './seatsContainer';
 
 export default function App() {
   const dispatch = useAppDispatch();
@@ -25,7 +25,7 @@ export default function App() {
   return (
     <div>
       <HeaderContainer/>
-      <Seats/>
+      <SeatsContainer/>
     </div>
   );
 }
