--- app-web/src/SelectTimeContainer.tsx
@@ -1,9 +1,0 @@
-import SelectTime from './SelectTime';
-
-export default function SelectTimeContainer({ seatNumber, onClick, onDelete, isMySeat }:any) {
-  return (
-    <>
-      <SelectTime seatNumber={seatNumber} onClick={onClick} onDelete={onDelete} isMySeat={isMySeat}/>
-    </>
-  );
-}
