--- app-web/src/admin.tsx
@@ -1,15 +1,0 @@
-import { Link } from 'react-router-dom';
-
-export default function Admin() {
-  return (
-    <section>
-      <div>
-        <Link to='/'>home</Link>
-      </div>
-      <h1>관리자 페이지</h1>
-      <br/>
-
-    </section>
-  );
-}
-
