+++ app-web/src/seatDetailModal.tsx
@@ -1,20 +1,13 @@
 import Modal from './components/Modal';
 import Detail from './components/details';
-import SelectTimeContainer from './SelectTimeContainer';
 
-
-export default function SeatDetailModal({ open, onClose, seatNumber, seatDetail, onClick, onDelete }:any) {
+export default function SeatDetailModal({ open, onClose, seatNumber, seatDetail }:any) {
   if (!seatDetail) return (
     <>
       <Modal open={open} onClose={onClose} >
         <Detail
           userName=''
           seatNumber={seatNumber}
-        />
-        <SelectTimeContainer
-          seatNumber={seatNumber}
-          onClick={onClick}
-          onDelete={onDelete}
         />
       </Modal>
     </>
@@ -27,11 +20,6 @@ export default function SeatDetailModal({ open, onClose, seatNumber, seatDetail,
         <Detail
           userName={userName}
           seatNumber={seatNumber}
-        />
-        <SelectTimeContainer
-          seatNumber={seatNumber}
-          onClick={onClick}
-          onDelete={onDelete}
         />
       </Modal>
     </>
